Rename home icon to house icon
